Add files via upload
Unzip the contents of the zip file to any folder.

This release includes 2 main uses. 
The first use is a bat file that will copy the icon hll.ico to a system directory (%ProgramData%\Hell Let Loose).  It will then create 6 steam direct launch shortcuts on your desktop, 1 for each EASY Company Server.  

To Run the bat file open the folder you unzipped the contents to and run (double click) HLL Shortcuts.bat.  You do not need administrative pillages.

If you choose to run the bat file, once the shortcuts are created, you may delete the extracted files.

The second use is if you prefer not to run the bat file.  
Included in the Zip file is a webpage Steam Launch.html.  Copy Steam Launch.html from the folder you extracted the zip file, and paste it on your desktop.  Double click on Steam Launch.html to open the file in your browser.  From there you can click on any of the links to direct connect to the EASY Company HLL servers.